Neo4j moved (#1152)
* Neo4j cleanup

* Neo4j cleanup

* Neo4j cleanup x2
