+++ spring-data-neo4j/src/main/java/com/baeldung/spring/data/neo4j/domain/Company.java
@@ -1,11 +1,8 @@
-package com.baeldung.graph;
+package com.baeldung.spring.data.neo4j.domain;
 
 import org.neo4j.ogm.annotation.NodeEntity;
 import org.neo4j.ogm.annotation.Relationship;
 
-/**
- * @author Danil Kornishev (danil.kornishev@mastercard.com)
- */
 @NodeEntity
 public class Company {
     private Long id;
