+++ spring-data-neo4j/src/test/java/com/baeldung/neo4j/Neo4JServerTest.java
@@ -1,10 +1,11 @@
-package com.baeldung.graph;
+package com.baeldung.neo4j;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
 import java.sql.Statement;
 
+import org.junit.Assert;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.neo4j.driver.v1.AuthTokens;
@@ -12,7 +13,6 @@ import org.neo4j.driver.v1.Driver;
 import org.neo4j.driver.v1.GraphDatabase;
 import org.neo4j.driver.v1.Session;
 import org.neo4j.driver.v1.StatementResult;
-import org.testng.Assert;
 
 @Ignore
 public class Neo4JServerTest {
