+++ spring-data-neo4j/src/test/resources/logback.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+
+    <appender name="console" class="ch.qos.logback.core.ConsoleAppender">
+        <encoder>
+            <pattern>%d %5p %40.40c:%4L - %m%n</pattern>
+        </encoder>
+    </appender>
+
+    <logger name="org.neo4j.ogm" level="info"/>
+
+    <root level="warn">
+        <appender-ref ref="console"/>
+    </root>
+
+</configuration>
